feat: moving implementation to common-files
