+++ cli/lib/nf3.mjs
@@ -900,7 +900,7 @@ class Nf3 {
     };
     connection.onmessage = async message => {
       const msg = JSON.parse(message.data);
-      const { number, type, txDataToSign, block, transactions } = msg;
+      const { id, type, txDataToSign, block, transactions } = msg;
       logger.debug(`Proposer received websocket message of type ${type}`);
       if (type === 'block') {
         proposerQueue.push(async () => {
@@ -916,7 +916,7 @@ class Nf3 {
           }
         });
 
-        connection.send(number);
+        connection.send(id);
       }
       return null;
     };
