+++ nightfall-optimist/src/event-handlers/chain-reorg.mjs
@@ -66,7 +66,7 @@ import {
   deleteTreeByBlockNumberL2,
   deleteNullifiersForBlock,
 } from '../services/database.mjs';
-import { waitForContract } from './subscribe.mjs';
+import { waitForContract } from '../utils/index.mjs';
 
 const { STATE_CONTRACT_NAME } = config;
