+++ nightfall-optimist/src/index.mjs
@@ -1,14 +1,9 @@
 import logger from 'common-files/utils/logger.mjs';
 import { queueManager, queues, enqueueEvent } from 'common-files/utils/event-queue.mjs';
+import { NFWebsocket } from 'common-files/utils/websocket.mjs';
+import config from 'config';
 import app from './app.mjs';
-import {
-  startEventQueue,
-  subscribeToBlockAssembledWebSocketConnection,
-  subscribeToChallengeWebSocketConnection,
-  subscribeToInstantWithDrawalWebSocketConnection,
-  subscribeToProposedBlockWebSocketConnection,
-  eventHandlers,
-} from './event-handlers/index.mjs';
+import { startEventQueue, eventHandlers } from './event-handlers/index.mjs';
 import Proposer from './classes/proposer.mjs';
 import {
   setBlockAssembledWebSocketConnection,
@@ -20,15 +15,27 @@ import { setInstantWithdrawalWebSocketConnection } from './services/instant-with
 import { setProposer } from './routes/proposer.mjs';
 import { setBlockProposedWebSocketConnection } from './event-handlers/block-proposed.mjs';
 
+const { WEBSOCKET_PORT, WEBSOCKET_PING_TIME } = config;
+
 const main = async () => {
   try {
     const proposer = new Proposer();
     setProposer(proposer); // passes the proposer instance int the proposer routes
     // subscribe to WebSocket events first
-    await subscribeToBlockAssembledWebSocketConnection(setBlockAssembledWebSocketConnection);
-    await subscribeToChallengeWebSocketConnection(setChallengeWebSocketConnection);
-    await subscribeToInstantWithDrawalWebSocketConnection(setInstantWithdrawalWebSocketConnection);
-    await subscribeToProposedBlockWebSocketConnection(setBlockProposedWebSocketConnection);
+
+    const ws = new NFWebsocket({ port: WEBSOCKET_PORT, pingTime: WEBSOCKET_PING_TIME });
+
+    ws.subscribe({ topic: 'challenge', socketName: 'challenge' }, setChallengeWebSocketConnection);
+    ws.subscribe({ topic: 'blocks', socketName: 'proposer' }, setBlockAssembledWebSocketConnection);
+    ws.subscribe(
+      { topic: 'instant', socketName: 'liquidity provider' },
+      setInstantWithdrawalWebSocketConnection,
+    );
+    ws.subscribe(
+      { topic: 'sync', socketName: 'publisher', filter: 'type' },
+      setBlockProposedWebSocketConnection,
+    );
+
     // start the event queue
     await startEventQueue(queueManager, eventHandlers, proposer);
     // enqueue the block-assembler every time the queue becomes empty
