+++ nightfall-optimist/src/services/transaction-checker.mjs
@@ -9,7 +9,7 @@ import config from 'config';
 import gen from 'general-number';
 import logger from 'common-files/utils/logger.mjs';
 import { Transaction, VerificationKey, Proof, TransactionError } from '../classes/index.mjs';
-import { waitForContract } from '../event-handlers/subscribe.mjs';
+import { waitForContract } from '../utils/index.mjs';
 import { getBlockByBlockNumberL2 } from './database.mjs';
 import verify from './verify.mjs';
